refactoring: feature-toggling azure cli parsing/cloudshell auth
